# cdv-api/docs/swagger.yaml
basePath: /cdv-api
definitions:
  internal_cabins.CabinCardResponse:
    properties:
      address:
        type: string
      capacity:
        type: integer
      host_id:
        type: integer
      host_name:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  internal_cabins.CabinCardsListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_cabins.CabinCardResponse'
        type: array
      total:
        type: integer
    type: object
  internal_cabins.CabinResponse:
    properties:
      address:
        type: string
      capacity:
        type: integer
      commission_id:
        type: integer
      description:
        type: string
      host_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      rules:
        type: string
    type: object
  internal_cabins.CabinsListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_cabins.CabinResponse'
        type: array
      total:
        type: integer
    type: object
  internal_cabins.CreateCabinRequest:
    properties:
      address:
        type: string
      capacity:
        type: integer
      commission_id:
        type: integer
      description:
        type: string
      host_id:
        type: integer
      name:
        type: string
      price:
        type: number
      rules:
        type: string
    type: object
  internal_cabins.CreateCabinResponse:
    properties:
      cabin_id:
        type: integer
      message:
        type: string
    type: object
  internal_cabins.DeleteCabinResponse:
    properties:
      message:
        type: string
      rows_affected:
        type: integer
    type: object
  internal_cabins.UpdateCabinRequest:
    properties:
      address:
        type: string
      capacity:
        type: integer
      description:
        type: string
      name:
        type: string
      price:
        type: number
      rules:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Documentación de los endpoints de la API de Casa Del Valle.
  title: Casa Del Valle API
  version: "1.0"
paths:
  /cabins:
    get:
      description: Devuelve el listado completo de cabañas.
      produces:
      - application/json
      responses:
        "200":
          description: Lista de cabañas
          schema:
            $ref: '#/definitions/internal_cabins.CabinsListResponse'
        "500":
          description: Error al obtener las cabañas
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Listar cabañas
      tags:
      - Cabañas
    post:
      consumes:
      - application/json
      description: Crea una cabaña. Para anfitriones el id_anfitrion se toma del token;
        para administradores puede indicarse. Requiere token.
      parameters:
      - description: Datos de la cabaña
        in: body
        name: cabin
        required: true
        schema:
          $ref: '#/definitions/internal_cabins.CreateCabinRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Cabaña registrada exitosamente
          schema:
            $ref: '#/definitions/internal_cabins.CreateCabinResponse'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Sin permisos
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Registrar cabaña
      tags:
      - Cabañas
  /cabins/{id}:
    delete:
      description: Elimina una cabaña. Solo el dueño de la cabaña o un administrador.
        Requiere token.
      parameters:
      - description: ID de cabaña
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Cabaña eliminada exitosamente
          schema:
            $ref: '#/definitions/internal_cabins.DeleteCabinResponse'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Sin permisos
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Cabaña no encontrada
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar cabaña
      tags:
      - Cabañas
    get:
      description: Devuelve los detalles de la cabaña indicada. Requiere token.
      parameters:
      - description: ID de cabaña
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Detalle de la cabaña
          schema:
            $ref: '#/definitions/internal_cabins.CabinResponse'
        "400":
          description: ID de cabaña inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Cabaña no encontrada
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Obtener cabaña por ID
      tags:
      - Cabañas
    put:
      consumes:
      - application/json
      description: Actualiza los datos de una cabaña. Solo el dueño de la cabaña o
        un administrador. Requiere token.
      parameters:
      - description: ID de cabaña
        in: path
        name: id
        required: true
        type: integer
      - description: Datos a actualizar
        in: body
        name: cabin
        required: true
        schema:
          $ref: '#/definitions/internal_cabins.UpdateCabinRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cabaña actualizada exitosamente
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Datos inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Sin permisos
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Cabaña no encontrada
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar cabaña
      tags:
      - Cabañas
...
  /cabins/search:
    get:
      description: Busca cabañas combinando filtros opcionales por nombre, anfitrión,
        capacidad y precio. Requiere token.
      parameters:
      - description: Filtro por nombre
        in: query
        name: name
        type: string
      - description: Filtro por ID de anfitrión
        in: query
        name: host_id
        type: integer
      - description: Filtro por nombre de anfitrión
        in: query
        name: host_name
        type: string
      - description: Capacidad mínima
        in: query
        name: min_capacity
        type: integer
      - description: Capacidad máxima
        in: query
        name: max_capacity
        type: integer
      - description: Precio mínimo
        in: query
        name: min_price
        type: number
      - description: Precio máximo
        in: query
        name: max_price
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Resultados de la búsqueda
          schema:
            $ref: '#/definitions/internal_cabins.CabinCardsListResponse'
        "400":
          description: Parámetros inválidos
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Buscar cabañas
      tags:
      - Cabañas
  /cabins/user/{userId}:
    get:
      description: Devuelve las cabañas cuyo anfitrión es el usuario indicado. Solo
        el propio usuario o un administrador. Requiere token.
      parameters:
      - description: ID del anfitrión
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cabañas del usuario
          schema:
            $ref: '#/definitions/internal_cabins.CabinsListResponse'
        "400":
          description: ID de usuario inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Sin permisos
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cabañas de un usuario
      tags:
      - Cabañas
...
securityDefinitions:
  BearerAuth:
    description: 'Token JWT obtenido en el endpoint login. Enviar como: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"